Add new CPT for siblings.
This new CPT is essentially a copy of Warriors. But to keep the URL
structure the way we want, I decided to make this a separate CPT instead
of a taxonomy under warriors.

The admin link, however, is nested under Warriors to keep the admin
clean and to offer a bit of a relationshop between the two without
needing to set up a bunch of custom rewrite rules.